Feature/ Secure web-native communication via an authorization cookie
To ensure at any point in time only the web ssap may call into the native endpoint, the ssap will "authorize" the calls by sending an authorization cookie with which it has been initialized on app start. The value of the cookie will be dynamically generated on each app start.